+++ CHANGES.rst
@@ -9,7 +9,14 @@ Changes
 `Unreleased <https://github.com/Ouranosinc/Magpie/tree/master>`_ (latest)
 ------------------------------------------------------------------------------------
 
-* Nothing new for the moment.
+Features / Changes
+~~~~~~~~~~~~~~~~~~~~~
+* Add support of RESTful API endpoints (i.e.: ``ServiceAPI``) under ``ServiceGeoserver`` using ``Route`` resources
+  (fixes `#584 <https://github.com/Ouranosinc/Magpie/issues/584>`_).
+  Requires the `Service` to be configured either with the default ``configuration``,
+  or by explicitly setting ``api: true``. When a HTTP request is sent toward a `Service` typed ``ServiceGeoserver``,
+  any non-`OWS` request (i.e.: `WFS`, `WMS`, `WPS`) will default to the resolution handling of typical ``ServiceAPI``.
+  This can be used notably to access the ``/web`` and ``/ogc`` endpoints of a `GeoServer` instance.
 
 .. _changes_3.34.0: